Stress Tests added. 10x as many tests.

+++ Nasa Test Engine 1/src/main/java/nasaTestSuite/FrigiDriver.java
@@ -573,7 +573,7 @@ public class FrigiDriver extends AndroidDriver
 		}	
 
 		long stopTime = System.currentTimeMillis();
-		System.out.println("Seconds to tap by coordinates: " + ((stopTime - startTime)/1000));
+		System.out.println("Seconds to tap by coordinates: " + (((float)stopTime - (float)startTime)/1000f));
 		return elementLocation;
 	}
